Added open map
The map can now be opened for a wider view of the map for better locating of death marks.
Additional death marks were added across the entire map.
The inputs variable was moved to the player values script to ensure that only one instance existed and was enabled at a time for improved performance.